Feat/clubtoday

+++ src/domain/usecase/clubtoday/DeleteClubTodayUseCase.ts
@@ -1,3 +1,5 @@
+import { Code } from '@/common/code/Code';
+import { Exception } from '@/common/exception/Exception';
 import { IUseCase } from '@/common/usecase/IUseCase';
 import { IDeleteClubTodayPort } from '@/domain/port/clubtoday/IDeleteClubTodayPort';
 import { IClubTodayRepository } from '@/domain/repository/IClubTodayRepository';
@@ -11,7 +13,9 @@ export class DeleteClubTodayUseCase implements IUseCase<IDeleteClubTodayPort, vo
    * @step_1 port로 받아온 id값으로 클럽투데이를 삭제한다.
    * @returns void
    */
-  execute(port?: IDeleteClubTodayPort): Promise<void> {
+  async execute(port?: IDeleteClubTodayPort): Promise<void> {
+    const clubTodayExist = await this.clubTodayRepository.getClubTodayById(port.id);
+    if (!clubTodayExist) throw Exception.new({ code: Code.NOT_FOUND, overrideMessage: '없는 클럽투데이' });
     return this.clubTodayRepository.deleteClubTodayById(port.id);
   }
 }
